Remove compiler warnings.
git-svn-id: svn+ssh://lc-svn.llnl.gov/usr/gapps/emhm/svn/Cardioid/trunk@84 baecf98e-1e81-4516-8920-d4c889e67ea4

+++ src/Koradi.cc
@@ -337,8 +337,8 @@ void Koradi::printStatistics()
 {
    computeRadii();
 
-   int maxLoad = *max_element(load_.begin(), load_.end());
-   int minLoad = *min_element(load_.begin(), load_.end());
+   double maxLoad = *max_element(load_.begin(), load_.end());
+   double minLoad = *min_element(load_.begin(), load_.end());
    double maxRadius = *max_element(radii_.begin(), radii_.end());
    double minRadius = *min_element(radii_.begin(), radii_.end());
    double maxBias = *max_element(bias_.begin(), bias_.end());
